+++ configurer/core/models/__init__.py
@@ -1,6 +1,7 @@
 from .certificates_manager import CertsManager
 from .wazuh_components_config_manager import WazuhComponentConfigManager
 from .wazuh_config_mapping import (
+    WazuhAgentConfigMapping,
     WazuhConfigMapping,
     WazuhDashboardConfigMapping,
     WazuhIndexerConfigMapping,