+++ configurer/core/models/wazuh_config_mapping.py
@@ -45,7 +45,11 @@ class WazuhConfigMapping:
             if (file_content := file.get(type, None)) and file.get("path", None):
                 try:
                     content.append(
-                        {"path": Path(file["path"]), "keys": file_content["keys"], "values": file_content["values"]}
+                        {
+                            "path": Path(file["path"]),
+                            "keys": file_content["keys"],
+                            "values": file_content["values"],
+                        }
                     )
                 except KeyError as err:
                     raise KeyError(f"Missing 'keys' or 'values' key in '{type}' mapping file section: {err}") from err
@@ -63,4 +67,8 @@ class WazuhServerConfigMapping(WazuhConfigMapping):
 
 
 class WazuhDashboardConfigMapping(WazuhConfigMapping):
+    pass
+
+
+class WazuhAgentConfigMapping(WazuhConfigMapping):
     pass